feat: create script to transform json to csv
