+++ prisma/seed/seed.ts
@@ -1,0 +1,37 @@
+import { PrismaClient } from "@prisma/client"
+import { participants } from "./participants"
+
+const prisma = new PrismaClient()
+
+async function main() {
+  const event = await prisma.event.create({
+    data: {
+      name: "Code in the Dark 2023",
+      slug: "2023",
+    },
+  })
+
+  const promiseParticipants = participants.map((data) =>
+    prisma.user.create({
+      data: {
+        ...data,
+        play: {
+          create: {
+            eventSlug: event.slug,
+          },
+        },
+      },
+    }),
+  )
+  await Promise.all(promiseParticipants)
+}
+
+main()
+  .then(async () => {
+    await prisma.$disconnect()
+  })
+  .catch(async (e) => {
+    console.log(e)
+    await prisma.$disconnect()
+    process.exit(1)
+  })
